add services, fix entity relations

+++ bank-system/src/main/java/com/alekhnovich/bsu/fpmi2019/banksystem/respository/BankRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BankRepository extends JpaRepository<Bank,Long> {
+public interface BankRepository extends JpaRepository<Bank,Integer> {
     Bank getBankByBankName(String bankName);
     Bank getBankByAddress(String address);
 }
