+++ bank-system/src/main/java/com/alekhnovich/bsu/fpmi2019/banksystem/respository/SwiftRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SwiftRepository extends JpaRepository<SwiftDirectory,Long>{
+public interface SwiftRepository extends JpaRepository<SwiftDirectory,Integer>{
 }
