+++ bank-system/src/main/java/com/alekhnovich/bsu/fpmi2019/banksystem/services/BankAccountService.java
@@ -2,6 +2,6 @@ package com.alekhnovich.bsu.fpmi2019.banksystem.services;
 
 import com.alekhnovich.bsu.fpmi2019.banksystem.models.BankAccount;
 
-public interface BankAccountService extends BaseService<BankAccount,Long>{
+public interface BankAccountService extends BaseService<BankAccount,Integer>{
     BankAccount getBankAccountByNumber(String number);
 }
