+++ bank-system/src/main/java/com/alekhnovich/bsu/fpmi2019/banksystem/services/BankService.java
@@ -2,7 +2,7 @@ package com.alekhnovich.bsu.fpmi2019.banksystem.services;
 
 import com.alekhnovich.bsu.fpmi2019.banksystem.models.Bank;
 
-public interface BankService extends BaseService<Bank,Long> {
+public interface BankService extends BaseService<Bank,Integer> {
     Bank getBankByName(String name);
     Bank getBankByAddress(String address);
 }
